Refactor out high-res and purchase vuex states
