+++ src/components/ui-controls/CompletePurchaseButton.vue
@@ -55,7 +55,7 @@
   import { mapGetters, mapState } from 'vuex';
   import _ from 'lodash';
   import * as actions from '../../store/actions';
-  import * as mutations from '../../store/mutation-types';
+  import * as mutations from '../../store/mutation';
   import AddressIcon from './AddressIcon.vue';
 
   export default {
@@ -70,13 +70,20 @@
     computed: {
       ...mapGetters([
         'isKnownOrigin',
+      ]),
+      ...mapState([
+        'account',
+      ]),
+      ...mapGetters('purchase', [
         'assetPurchaseState',
         'isPurchaseTriggered',
         'isPurchaseStarted',
         'isPurchaseSuccessful',
         'isPurchaseFailed',
       ]),
-      ...mapState(['account', 'purchaseState']),
+      ...mapState('purchase', [
+        'purchaseState'
+      ]),
       soldAsFiat: function () {
         return this.asset.purchased === 2;
       },
@@ -93,17 +100,17 @@
       completePurchase: function () {
         console.log('Completing purchase', this.asset);
         this.$emit('purchaseInitiated', this.asset);
-        this.$store.dispatch(actions.PURCHASE_ASSET, this.asset);
+        this.$store.dispatch(`purchase/${actions.PURCHASE_ASSET}`, this.asset);
       },
       completeFiatPurchase: function () {
         console.log('Completing FIAT purchase', this.asset);
         this.$emit('purchaseInitiated', this.asset);
-        this.$store.dispatch(actions.PURCHASE_ASSET_WITH_FIAT, this.asset);
+        this.$store.dispatch(`purchase/${actions.PURCHASE_ASSET_WITH_FIAT}`, this.asset);
       },
       reverseFiatPurchase: function () {
         console.log('Reverse FIAT purchase', this.asset);
         this.$emit('purchaseInitiated', this.asset);
-        this.$store.dispatch(actions.REVERSE_PURCHASE_ASSET_WITH_FIAT, this.asset);
+        this.$store.dispatch(`purchase/${actions.REVERSE_PURCHASE_ASSET_WITH_FIAT}`, this.asset);
       }
     }
   };
