+++ src/store/mutation.js
@@ -4,14 +4,10 @@
 export const SET_CURRENT_NETWORK = 'SET_CURRENT_NETWORK';
 export const SET_ETHERSCAN_NETWORK = 'SET_ETHERSCAN_NETWORK';
 export const SET_WEB3 = 'SET_WEB3';
+export const SET_KODA_CONTRACT = 'SET_KODA_CONTRACT';
 
 // External API calls
 export const SET_USD_PRICE = 'SET_USD_PRICE';
-
-// High-res
-export const HIGH_RES_DOWNLOAD_SUCCESS = 'HIGH_RES_DOWNLOAD_SUCCESS';
-export const HIGH_RES_DOWNLOAD_FAILURE = 'HIGH_RES_DOWNLOAD_FAILURE';
-export const HIGH_RES_DOWNLOAD_TRIGGERED = 'HIGH_RES_DOWNLOAD_TRIGGERED';
 
 ///////////////////
 // Contract Data //
@@ -27,7 +23,12 @@ export const SET_COMMISSION_ADDRESSES = 'SET_COMMISSION_ADDRESSES';
 
 export const SET_TOTAL_PURCHASED = 'SET_TOTAL_PURCHASED';
 
-export const RESET_PURCHASE_STATE = 'RESET_PURCHASE_STATE';
+//////////////
+// High-res //
+//////////////
+export const HIGH_RES_DOWNLOAD_SUCCESS = 'HIGH_RES_DOWNLOAD_SUCCESS';
+export const HIGH_RES_DOWNLOAD_FAILURE = 'HIGH_RES_DOWNLOAD_FAILURE';
+export const HIGH_RES_DOWNLOAD_TRIGGERED = 'HIGH_RES_DOWNLOAD_TRIGGERED';
 
 ///////////////////
 // Purchase flow //
@@ -36,3 +37,4 @@ export const PURCHASE_TRIGGERED = 'PURCHASE_TRIGGERED'; // metamask opened
 export const PURCHASE_STARTED = 'PURCHASE_STARTED'; // metamask transaction submitted
 export const PURCHASE_FAILED = 'PURCHASE_FAILED'; // a failure
 export const PURCHASE_SUCCESSFUL = 'PURCHASE_SUCCESSFUL'; // transaction mined and Purchase event heard
+export const RESET_PURCHASE_STATE = 'RESET_PURCHASE_STATE';
